test(voice-identity): 角色生命周期测试移出 test_memory_zh_tw
这 5 个测试覆盖的是 character_runtime 对 voice_identity_runtime 的
register/unregister 接线和删除/初始化并发的 websocket_lock 事务，跟
"记忆繁体" 没有关系，只是当初借用了该文件里已有的 role_state 夹具。

移到 tests/unit/test_character_runtime_voice_identity.py，自带 import
和模块 docstring。纯搬家，测试体一行未改。
